@@ -12,12 +12,12 @@ use Becklyn\Ddd\Events\Domain\EventId;
  */
 abstract class <?= $class_name; ?> extends AbstractDomainEvent
 {
-    protected <?= $entity; ?>Id $<?= $strtocamel($entity); ?>Id;
-
-    public function __construct (EventId $id, \DateTimeImmutable $raisedTs, <?= $entity; ?>Id $<?= $strtocamel($entity); ?>Id)
-    {
+    public function __construct (
+        EventId $id,
+        \DateTimeImmutable $raisedTs,
+        protected <?= $entity; ?>Id $<?= $strtocamel($entity); ?>Id,
+    ){
         parent::__construct($id, $raisedTs);
-        $this-><?= $strtocamel($entity); ?>Id = $<?= $strtocamel($entity); ?>Id;
     }
 
     public function aggregateId () : <?= $entity; ?>Id
